apis/requirements.txt
dnspython==2.4.2
Flask==3.0.0
Flask-Cors==4.0.0
jwt==1.3.1
pymongo==4.5.0
python-dotenv==1.0.0
Werkzeug==3.0.0
